+++ CHANGES.md
@@ -60,6 +60,7 @@ Changes in 12.5
 - JS remove inline code from program in Profiles.php & Profiles.js
 - JS remove inline code from program in StudentLabels.fnc.php & StudentLabels.js
 - JS remove inline code from BottomButtonBackUpdate() in Bottom.fnc.php & BottomButtonBackUpdate.js
+- JS remove inline code from ChartjsChart() in Charts.fnc.php & ChartjsChart.js
 
 Changes in 12.4.3
 -----------------